test: auth-service (#25)
