+++ auth/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "ts-node-dev src/index.ts",
-    "prisma:migrate": "sleep 10 && npx prisma migrate dev"
+    "prisma:migrate": "sleep 10 && npx prisma migrate dev",
+    "test": "jest --watchAll --no-cache"
   },
   "keywords": [],
   "author": "",
@@ -29,8 +30,13 @@
   "devDependencies": {
     "@types/bcryptjs": "^2.4.2",
     "@types/cookie-session": "^2.0.44",
+    "@types/jest": "^29.5.2",
     "@types/jsonwebtoken": "^9.0.2",
     "@types/nodemailer": "^6.4.8",
-    "prisma": "^4.15.0"
+    "@types/supertest": "^2.0.12",
+    "jest": "^29.5.0",
+    "prisma": "^4.15.0",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.0"
   }
 }
