+++ auth/src/index.ts
@@ -1,35 +1,4 @@
-import express from "express";
-import "express-async-errors";
-import { json } from "body-parser";
-import cookieSession from "cookie-session";
-
-import { currentUserRouter } from "./routes/current-user";
-import { signupRouter } from "./routes/signup";
-import { signinRouter } from "./routes/signin";
-import { signoutRouter } from "./routes/signout";
-import { errorHandler } from "./middlewares/error-handler";
-import { NotFoundError } from "./errors/not-found-error";
-
-const app = express();
-app.set("trust proxy", true);
-app.use(json());
-app.use(
-  cookieSession({
-    signed: false,
-    secure: true,
-  })
-);
-
-app.use(signupRouter);
-app.use(signinRouter);
-app.use(signoutRouter);
-app.use(currentUserRouter);
-
-app.get("*", async () => {
-  throw new NotFoundError();
-});
-
-app.use(errorHandler);
+import app from "./app";
 
 app.listen(4000, () => {
   console.log("서버가 포트번호 4000에서 정상 작동중입니다.");
