+++ auth/src/routes/signup.ts
@@ -13,7 +13,7 @@ router.post(
       .trim()
       .isLength({ min: 4, max: 20 })
       .withMessage("비밀번호는 4자리 이상 20자리 이하여야 합니다."),
-    body("username")
+    body("name")
       .trim()
       .isLength({ min: 4, max: 10 })
       .withMessage("닉네임은 4자리 이상 10자리 이하여야 합니다."),
